backend-python/requirements.txt
Flask
flask-cors
google-cloud-storage
google-cloud-speech
firebase-admin
google-generativeai
python-dotenv
